move events from data_models to audit module
