+++ src/dbally/audit/event_tracker.py
@@ -3,7 +3,7 @@ from typing import AsyncIterator, Dict, List, Optional, Union
 
 from dbally.audit.event_handlers.base import EventHandler
 from dbally.audit.event_span import EventSpan
-from dbally.data_models.audit import LLMEvent, RequestEnd, RequestStart, SimilarityEvent
+from dbally.audit.events import LLMEvent, RequestEnd, RequestStart, SimilarityEvent
 
 
 class EventTracker:
